requirements.txt
streamlit>=1.24.0
snowflake-snowpark-python>=1.5.1
pandas>=2.0.0
python-dotenv>=1.0.0 